Changed order book algorithm, untested size relations.

+++ schema/protocol.capnp
@@ -2,15 +2,15 @@
 
 struct OrderMsg {
     buy @0 :Bool;
-    volume @1 :Int32;
+    volume @1 :UInt32;
     limitprice :union {
         none @2 :Void;
-        some @3 :Int32;
+        some @3 :Float32;
     }
     condition :union {
         unconditional @4 :Void;
         stoporder :group {
-            stop @5 :Int32;
+            stop @5 :Float32;
         }
     }
     assetname @6 :UInt16;
@@ -18,8 +18,8 @@ struct OrderMsg {
 
 
 struct VolumeTradedAtPrice {
-    volume @0 :Int32;
-    price @1 :Int32;
+    volume @0 :UInt32;
+    price @1 :Float32;
 }
